feat: 간호기록 삭제 시 사용자 확인 절차 추가

+++ src/routes/Main/DisplayInformation/NursingRecord/ActionButtons.tsx
@@ -20,6 +20,9 @@ const ActionButtons = ({ refetch, ...request }: Props) => {
   const handleClose = () => setAnchorEl(null);
 
   const onDelete = () => {
+    const isConfirm = window.confirm('간호기록을 삭제하시겠습니까?');
+    if (!isConfirm) return handleClose();
+
     deleteNursingRecord(request)
       .then(() => {
         refetch();
